+++ src/Interface/IModulaOptions.ts
@@ -7,6 +7,7 @@ export default interface IModulaOptions
     routes?: IRoute[];
     components?: IComponentDefinition[];
     template?: new () => Component;
+    guestTemplate?: new () => Component;
     pageNotFound?: new () => Component;
     userProp?: string;
     authRedirectPath?: string;